Add armi.reactor.grids.hexagonal
